RocketMQ 任意延时消息优化基本功能

+++ client/src/main/java/org/apache/rocketmq/client/schedule/ScheduleMQProducer.java
@@ -43,6 +43,6 @@ public interface ScheduleMQProducer extends MQProducer {
      * @throws MQBrokerException
      * @throws InterruptedException
      */
-    SendResult send(final Message msg, final long timeout, int delayTime, TimeUnit timeUnit) throws MQClientException,
+    SendResult send(final Message msg, final long timeout, long delayTime, TimeUnit timeUnit) throws MQClientException,
             RemotingException, MQBrokerException, InterruptedException;
 }
